# docs/logos/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Logos API",
    "version": "1.0.0",
    "description": "Versioned manuscripts for Chronos books: ordered volumes, ordered sections, and one structured rich-text document per section.\n\nPaths are relative to the service mount. In the combined deployment that is `/logos`; running Logos standalone it is `/`.\n\nEvery mutation of an existing resource requires an `If-Match` header carrying the `rev` from its last read. Unlike Chronos and Prithvi, Logos refuses an unconditional write: a lost update to prose cannot be reconstructed."
  },
  "servers": [
    { "url": "/logos", "description": "Combined single-origin deployment" },
    { "url": "/", "description": "Standalone deployment" }
  ],
  "security": [{ "sessionCookie": [] }],
...
  "paths": {
...
    "/books": {
      "get": {
        "tags": ["manuscripts"],
        "summary": "Every Chronos book you can read, and whether it has a manuscript",
        "description": "Counts only. Word totals need the prose itself, so they are on the manuscript and report reads instead.",
        "responses": {
          "200": {
            "description": "The shelf.",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/BookList" }
              }
            }
          },
          "401": { "$ref": "#/components/responses/Unauthorized" }
        }
      }
    },
    "/books/{book}": {
      "get": {
        "tags": ["manuscripts"],
        "summary": "The manuscript outline for one book",
        "description": "A book with no manuscript yet reads as `rev` 0 with an empty volume list, and carries no `ETag`.",
        "parameters": [{ "$ref": "#/components/parameters/Book" }],
        "responses": {
          "200": {
            "description": "The manuscript.",
            "headers": { "ETag": { "$ref": "#/components/headers/ETag" } },
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/Manuscript" }
              }
            }
          },
          "401": { "$ref": "#/components/responses/Unauthorized" },
          "403": { "$ref": "#/components/responses/Forbidden" },
          "404": { "$ref": "#/components/responses/NotFound" }
        }
      },
      "delete": {
        "tags": ["manuscripts"],
        "summary": "Delete a manuscript and its retained history",
        "description": "The only hard delete Logos offers, and what must happen before the Chronos book can be removed. Refused with `409 CASCADE_REQUIRED` while it still holds volumes or sections, unless `cascade=true`.",
        "parameters": [
          { "$ref": "#/components/parameters/Book" },
          { "$ref": "#/components/parameters/IfMatch" },
          { "$ref": "#/components/parameters/Cascade" }
        ],
        "responses": {
          "204": { "description": "Gone, history included." },
          "400": { "$ref": "#/components/responses/BadRequest" },
          "401": { "$ref": "#/components/responses/Unauthorized" },
          "403": { "$ref": "#/components/responses/Forbidden" },
          "404": { "$ref": "#/components/responses/NotFound" },
          "409": { "$ref": "#/components/responses/Conflict" },
          "428": { "$ref": "#/components/responses/PreconditionRequired" }
        }
      }
    },
    "/books/{book}/report": {
      "get": {
        "tags": ["manuscripts"],
        "summary": "Progress across a book, and every reference that no longer lands",
        "parameters": [{ "$ref": "#/components/parameters/Book" }],
        "responses": {
          "200": {
            "description": "The report.",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/Report" }
              }
            }
          },
          "401": { "$ref": "#/components/responses/Unauthorized" },
          "403": { "$ref": "#/components/responses/Forbidden" },
          "404": { "$ref": "#/components/responses/NotFound" }
        }
      }
    },
    "/books/{book}/volume-order": {
      "put": {
        "tags": ["manuscripts"],
        "summary": "Rearrange the volumes",
        "description": "The body must list every current volume id exactly once. `If-Match` is the manuscript's revision, not a volume's.",
        "parameters": [
          { "$ref": "#/components/parameters/Book" },
          { "$ref": "#/components/parameters/IfMatch" }
        ],
        "requestBody": { "$ref": "#/components/requestBodies/VolumeOrder" },
        "responses": {
          "200": {
            "description": "The manuscript, renumbered.",
            "headers": { "ETag": { "$ref": "#/components/headers/ETag" } },
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/Manuscript" }
              }
            }
          },
          "400": { "$ref": "#/components/responses/BadRequest" },
          "401": { "$ref": "#/components/responses/Unauthorized" },
          "403": { "$ref": "#/components/responses/Forbidden" },
          "404": { "$ref": "#/components/responses/NotFound" },
          "409": { "$ref": "#/components/responses/Conflict" },
          "428": { "$ref": "#/components/responses/PreconditionRequired" }
        }
      }
    },
    "/books/{book}/volumes/{volume}": {
      "post": {
        "tags": ["volumes"],
        "summary": "Create a volume",
        "description": "The first volume creates the manuscript. New volumes go last; use volume-order to move them.",
        "parameters": [
          { "$ref": "#/components/parameters/Book" },
          { "$ref": "#/components/parameters/Volume" }
        ],
        "requestBody": { "$ref": "#/components/requestBodies/VolumeBody" },
        "responses": {
          "201": {
            "description": "Created.",
            "headers": { "ETag": { "$ref": "#/components/headers/ETag" } },
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/Volume" }
              }
            }
          },
          "400": { "$ref": "#/components/responses/BadRequest" },
          "401": { "$ref": "#/components/responses/Unauthorized" },
          "403": { "$ref": "#/components/responses/Forbidden" },
          "404": { "$ref": "#/components/responses/NotFound" },
          "409": { "$ref": "#/components/responses/Conflict" }
        }
      },
      "get": {
        "tags": ["volumes"],
        "summary": "One volume and its section list",
        "description": "Sections carry their counts but not their prose. Use the manuscript read for the text.",
        "parameters": [
          { "$ref": "#/components/parameters/Book" },
          { "$ref": "#/components/parameters/Volume" }
        ],
        "responses": {
          "200": {
            "description": "The volume.",
            "headers": { "ETag": { "$ref": "#/components/headers/ETag" } },
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/Volume" }
              }
            }
          },
          "401": { "$ref": "#/components/responses/Unauthorized" },
          "403": { "$ref": "#/components/responses/Forbidden" },
          "404": { "$ref": "#/components/responses/NotFound" }
        }
      },
      "put": {
        "tags": ["volumes"],
        "summary": "Replace a volume's title and overview",
        "description": "Section order belongs to the volume and is not part of this body; use section-order to change it.",
        "parameters": [
          { "$ref": "#/components/parameters/Book" },
          { "$ref": "#/components/parameters/Volume" },
          { "$ref": "#/components/parameters/IfMatch" }
        ],
        "requestBody": { "$ref": "#/components/requestBodies/VolumeBody" },
        "responses": {
          "200": {
            "description": "The volume.",
            "headers": { "ETag": { "$ref": "#/components/headers/ETag" } },
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/Volume" }
              }
            }
          },
          "400": { "$ref": "#/components/responses/BadRequest" },
          "401": { "$ref": "#/components/responses/Unauthorized" },
          "403": { "$ref": "#/components/responses/Forbidden" },
          "404": { "$ref": "#/components/responses/NotFound" },
          "409": { "$ref": "#/components/responses/Conflict" },
          "428": { "$ref": "#/components/responses/PreconditionRequired" }
        }
      },
      "delete": {
        "tags": ["volumes"],
        "summary": "Delete a volume",
        "description": "Refused with `409 CASCADE_REQUIRED` while it still holds sections, unless `cascade=true`.",
        "parameters": [
          { "$ref": "#/components/parameters/Book" },
          { "$ref": "#/components/parameters/Volume" },
          { "$ref": "#/components/parameters/IfMatch" },
          { "$ref": "#/components/parameters/Cascade" }
        ],
        "responses": {
          "204": { "description": "Gone." },
          "400": { "$ref": "#/components/responses/BadRequest" },
          "401": { "$ref": "#/components/responses/Unauthorized" },
          "403": { "$ref": "#/components/responses/Forbidden" },
          "404": { "$ref": "#/components/responses/NotFound" },
          "409": { "$ref": "#/components/responses/Conflict" },
          "428": { "$ref": "#/components/responses/PreconditionRequired" }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "sessionCookie": {
        "type": "apiKey",
        "in": "cookie",
        "name": "session",
        "description": "The shared login cookie. One session covers all four services."
      }
    },
    "headers": {
      "ETag": {
        "description": "The resource's current revision, quoted. Send it back as `If-Match`.",
        "schema": { "type": "string" }
      }
    },
    "parameters": {
      "Book": {
        "name": "book",
        "in": "path",
        "required": true,
        "description": "The Chronos book id.",
        "schema": { "type": "string" }
      },
      "Volume": {
        "name": "volume",
        "in": "path",
        "required": true,
        "schema": { "type": "string" }
      },
...
      "IfMatch": {
        "name": "If-Match",
        "in": "header",
        "required": true,
        "description": "The revision the caller read. Required: Logos refuses unconditional writes.",
        "schema": { "type": "string" }
      },
      "Cascade": {
        "name": "cascade",
        "in": "query",
        "required": false,
        "description": "Set `true` to accept deleting the prose this resource still holds.",
        "schema": { "type": "boolean" }
      }
    },
    "requestBodies": {
      "VolumeBody": {
        "required": true,
        "content": {
          "application/json": {
            "schema": { "$ref": "#/components/schemas/VolumeBody" }
          }
        }
      },
...
      "VolumeOrder": {
        "required": true,
        "content": {
          "application/json": {
            "schema": { "$ref": "#/components/schemas/VolumeOrder" }
          }
        }
      },
...
    },
    "responses": {
      "BadRequest": {
        "description": "A malformed identifier, body, document, ordering or precondition.",
        "content": {
          "application/json": { "schema": { "$ref": "#/components/schemas/Error" } }
        }
      },
      "Unauthorized": {
        "description": "Not signed in.",
        "content": {
          "application/json": { "schema": { "$ref": "#/components/schemas/Error" } }
        }
      },
      "Forbidden": {
        "description": "No matching grant on the Chronos book. Returned before existence is checked, so it never reveals whether the book is there.",
        "content": {
          "application/json": { "schema": { "$ref": "#/components/schemas/Error" } }
        }
      },
      "NotFound": {
        "description": "The Chronos book, manuscript, volume or section does not exist.",
        "content": {
          "application/json": { "schema": { "$ref": "#/components/schemas/Error" } }
        }
      },
      "Conflict": {
        "description": "A stale revision, an id already taken, a second prologue/epilogue/glossary, or a delete that would take prose with it.",
        "content": {
          "application/json": { "schema": { "$ref": "#/components/schemas/Error" } }
        }
      },
...
      "PreconditionRequired": {
        "description": "The mutation carried no `If-Match`.",
        "content": {
          "application/json": { "schema": { "$ref": "#/components/schemas/Error" } }
        }
      }
    },
    "schemas": {
...
      "Error": {
        "type": "object",
        "required": ["error"],
        "properties": {
          "error": { "type": "string", "description": "Human-readable." },
          "code": {
            "type": "string",
            "description": "Stable machine code.",
            "enum": [
              "ALREADY_EXISTS",
              "BOOK_NOT_FOUND",
              "CASCADE_REQUIRED",
              "CHRONOS_EVENT_NOT_FOUND",
              "FORBIDDEN",
              "INVALID_DOCUMENT",
              "INVALID_IDENTIFIER",
              "INVALID_ORDER",
              "INVALID_REVISION",
              "INVALID_SECTION",
              "INVALID_VOLUME",
              "LOGOS_ERROR",
              "MANUSCRIPT_NOT_FOUND",
              "PRECONDITION_REQUIRED",
              "REVISION_CONFLICT",
              "REVISION_NOT_RETAINED",
              "SECTION_KIND_IN_USE",
              "SECTION_NOT_FOUND",
              "VOLUME_NOT_FOUND"
            ]
          },
          "evidence": {
            "type": "object",
            "description": "What the refusal was about, for a caller to act on."
          }
        },
        "example": {
          "error": "One or more referenced Chronos events do not exist in this book.",
          "code": "CHRONOS_EVENT_NOT_FOUND",
          "evidence": { "events": ["not-a-scene"] }
        }
      },
      "ArticleRef": {
        "type": "object",
        "required": ["database", "collection", "id"],
        "properties": {
          "database": { "type": "string", "description": "The Akasha world." },
          "collection": { "type": "string" },
          "id": { "type": "string" }
        }
      },
...
      "Permissions": {
        "type": "object",
        "required": ["write", "delete"],
        "properties": {
          "write": { "type": "boolean" },
          "delete": { "type": "boolean" }
        }
      },
      "BookList": {
        "type": "object",
        "required": ["books"],
        "properties": {
          "books": {
            "type": "array",
            "items": {
              "type": "object",
              "required": [
                "book",
                "title",
                "has_manuscript",
                "volume_count",
                "_links"
              ],
              "properties": {
                "book": { "type": "string" },
                "title": { "type": "string", "nullable": true },
                "has_manuscript": { "type": "boolean" },
                "volume_count": { "type": "integer" },
                "_links": { "type": "object" }
              }
            }
          }
        },
        "example": {
          "books": [
            {
              "book": "ember-pact",
              "title": "The Ember Pact",
              "has_manuscript": true,
              "volume_count": 1,
              "_links": { "self": "/books/ember-pact" }
            }
          ]
        }
      },
...
      "Volume": {
        "type": "object",
        "required": [
          "book",
          "id",
          "number",
          "title",
          "overview",
          "sections",
          "section_count",
          "word_count",
          "_links",
          "rev",
          "created_by",
          "updated_by",
          "updated_at"
        ],
        "properties": {
          "book": { "type": "string" },
          "id": { "type": "string" },
          "number": { "type": "integer", "description": "Derived from volume order." },
          "title": { "type": "string" },
          "overview": { "type": "string" },
          "sections": {
            "type": "array",
            "items": { "$ref": "#/components/schemas/SectionSummary" }
          },
          "section_count": { "type": "integer" },
          "word_count": { "type": "integer" },
          "_links": { "type": "object" },
          "rev": { "type": "integer" },
          "created_by": { "type": "string", "nullable": true },
          "updated_by": { "type": "string", "nullable": true },
          "updated_at": { "type": "string", "nullable": true },
          "permissions": { "$ref": "#/components/schemas/Permissions" }
        },
        "example": {
          "book": "ember-pact",
          "id": "the-ember-pact",
          "number": 1,
          "title": "The Ember Pact",
          "overview": "Volume one.",
          "sections": [],
          "section_count": 0,
          "word_count": 0,
          "_links": {
            "self": "/books/ember-pact/volumes/the-ember-pact",
            "manuscript": "/books/ember-pact/volumes/the-ember-pact/manuscript",
            "section_order": "/books/ember-pact/volumes/the-ember-pact/section-order"
          },
          "rev": 1,
          "created_by": "mara",
          "updated_by": "mara",
          "updated_at": "2026-01-01T00:00:00+00:00"
        }
      },
...
      "Manuscript": {
        "type": "object",
        "required": [
          "book",
          "title",
          "overview",
          "world",
          "volumes",
          "volume_count",
          "word_count",
          "_links",
          "rev",
          "created_by",
          "updated_by",
          "updated_at"
        ],
        "properties": {
          "book": { "type": "string" },
          "title": { "type": "string", "nullable": true, "description": "Read from Chronos, never copied." },
          "overview": { "type": "string" },
          "world": { "type": "string", "nullable": true },
          "volumes": {
            "type": "array",
            "items": { "$ref": "#/components/schemas/Volume" }
          },
          "volume_count": { "type": "integer" },
          "word_count": { "type": "integer" },
          "_links": { "type": "object" },
          "rev": { "type": "integer", "description": "0 when no manuscript exists yet." },
          "created_by": { "type": "string", "nullable": true },
          "updated_by": { "type": "string", "nullable": true },
          "updated_at": { "type": "string", "nullable": true },
          "permissions": { "$ref": "#/components/schemas/Permissions" }
        },
        "example": {
          "book": "ember-pact",
          "title": "The Ember Pact",
          "overview": "",
          "world": "ember",
          "volumes": [],
          "volume_count": 0,
          "word_count": 0,
          "_links": {
            "self": "/books/ember-pact",
            "volume_order": "/books/ember-pact/volume-order",
            "report": "/books/ember-pact/report"
          },
          "rev": 0,
          "created_by": null,
          "updated_by": null,
          "updated_at": null,
          "permissions": { "write": true, "delete": true }
        }
      },
...
      "Report": {
        "type": "object",
        "required": [
          "book",
          "volume_count",
          "section_count",
          "word_count",
          "sections_with_missing_refs"
        ],
        "properties": {
          "book": { "type": "string" },
          "volume_count": { "type": "integer" },
          "section_count": { "type": "integer" },
          "word_count": { "type": "integer" },
          "sections_with_missing_refs": {
            "type": "array",
            "items": {
              "type": "object",
              "required": ["volume", "section", "missing_refs"],
              "properties": {
                "volume": { "type": "string" },
                "section": { "type": "string" },
                "missing_refs": {
                  "type": "array",
                  "items": { "$ref": "#/components/schemas/ArticleRef" }
                }
              }
            }
          }
        },
        "example": {
          "book": "ember-pact",
          "volume_count": 1,
          "section_count": 2,
          "word_count": 1840,
          "sections_with_missing_refs": [
            {
              "volume": "the-ember-pact",
              "section": "the-broken-gate",
              "missing_refs": [
                {
                  "database": "ember",
                  "collection": "characters",
                  "id": "lyra"
                }
              ]
            }
          ]
        }
      },
      "VolumeBody": {
        "type": "object",
        "required": ["title"],
        "additionalProperties": false,
        "properties": {
          "title": { "type": "string", "maxLength": 300 },
          "overview": { "type": "string", "maxLength": 10000 }
        },
        "example": { "title": "The Ember Pact", "overview": "Volume one." }
      },
...
      "VolumeOrder": {
        "type": "object",
        "required": ["volumes"],
        "additionalProperties": false,
        "properties": {
          "volumes": { "type": "array", "items": { "type": "string" } }
        },
        "example": { "volumes": ["the-ember-pact", "the-second-pact"] }
      },
...
    }
  }
}